ITEP-20887: Fix path traversal issues in manager (#71)
In model_directory_view.py, it was possible to create folders outside
the intended location as the `www-data` user by sending
specially-crafted API requests. Client-side validation was in place, but
the backend code was trusting user-supplied paths naively.

This patch adds a new path joining method which ensures the result path
is still contained within the model root folder, as intended.

Fixes ITEP-20887.

Signed-off-by: Steve Miller <steve.miller@intel.com>
Co-authored-by: Sarat Poluri <sarat.chandra.poluri@intel.com>