ITEP-68793: Support Volumetric Analytics in Scene Controller (#148)
This PR enables support for Volumetric Analytics for Regions of
Interest. That includes:
1. Detect intersection of the ROI and object mesh.
2. Region of interest can be basic or volumetric.
3. Supports specifying a buffer around the region of interest and
visualization of Buffer.
4. Supports both Convex and Concave polygons.
5. Ability to define buffer around object in object library.
6. Documentation on how to create spatial analytics and create objects
in object library.

TODO:
1. Allow updating ROI settings from 3D UI - P3.
2. Support DB upgrade to add latest fields - P1. Will get dealt with as
part of the stable release to next stable release migration fix. Won't
include pre-PR to post-PR migration.

How to test:
Create a region of interest. Try different shapes but not complex
polygons i.e no intersecting edges.
Observe the behavior of when the region event is triggered. By default,
it is when the object center is inside the polygon.
Update the ROI to enable "volumetric"
Observe when the region event is triggered. When volumetric, the events
are triggered as soon as the intersection between meshes is observed
(check in 3D UI to understand what the object mesh and ROI mesh look
like). Entry is detected when the first intersection happens, exit is
triggered when the meshes are no longer intersecting.
Increase the buffer to observe the intersection behavior change.
Change height to see it persists in 3D UI. (Height will factor in when
we detect intersection percentage and not just boolean intersection).

Go to Object Library.
Change the object properties to add buffer size and save changes.
Go to 3D UI and verify that the object size has been adjusted based on
the buffer size.
Consequently, verify the behavior of volume intersection is also using
the modified object size.