ITEP-72407: support editing ROI from 3D UI (#183)
Regions of interest can be edited in 3D UI.

Not covered:
Updating buffer_size and name will not render dynamically. 3D UI page
needs to be refreshed to visualize the new buffer_size and name
settings.

How to verify:
Create a region of interest via 2D UI
Go to 3D UI of the same scene
Edit settings of the created region like: volumetric, height,
buffer_size and name.
Click save. Refresh to observe the changes are reflected.

Click on Delete in 3D UI.
Observe that the ROI and the controls are deleted.