ITEP-73519 fix: docker compose OOB demo no video input with `dlstreamer-pipeline-server` enabled (#262)
This fixes a bug of missing input video files in `sample_data` volume
when `DLS=1`. The bug originates from invalid order of `sample_data`
volume creation and input videos conversion in `demo` make target. As a
result, when deploying the OOB demo for the first time on a clean
repository, the `sample_data` volume was created before the videos were
converted and DLS couldn't find the input videos in the volume mounted
to the container.

Changes introduced:

- The video conversion code was extracted from the `demo` target to a
newly added make target `convert-dls-videos` which converts the input
videos to `.ts` format if `DLS=1`, otherwise does nothing.
- `convert-dls-videos` make target was made a dependency of
`init-sample-data` make target which creates the volume.

Additional change: limited debug output when creating the `sample_data`
volume.