ITEP-72774 - Add Apache timeout to prevent long delays when WebSocket backend doesn't listen on port (#265)
The frontend code attempts to connect to both WebSocket ports - secure
and insecure (1884 and 9001) - using Promises and awaits **both**
responses.

In normal environments (e.g., Docker), if no listener is present, the
connection is refused immediately with ECONNREFUSED. No retries are
made, and the Promises resolve quickly.

In Kubernetes, however, the behavior changes. The connection is routed
to the Service IP by kube-proxy, which forwards it to one of the
matching Pod IPs. If the container isn't actually listening on the
specified port, the kernel might silently drop the packet instead of
refusing the connection. `kube-proxy` does not validate whether the port
is open.

As a result, the client's TCP stack retries the connection multiple
times with exponential backoff, leading to a long delay before the
connection fails. The number of retries is controlled by the
`tcp_syn_retries` sysctl setting.

Default value is 6, which corresponds to 63 seconds till the last
retransmission with the current initial RTO of 1 second. With this the
final timeout for an active TCP connection attempt will happen after
127seconds.

Only after this timeout do both Promises resolve, allowing the secure
WebSocket connection to proceed. Until then, the cameras remain offline
in the web UI.