Update `dlstreamer-pipeline-server` version tag to `3.1.0` (#317)
Update `dlstreamer-pipeline-server` version tag to `3.1.0` in Docker and
Kubernetes deployments to use the latest public image, including tests
and docs.

Using lazy import of `scipy` for 3D detector as base image lacks `scipy`
installed.
Changes the healthchecks from using `wget` to `curl` as `wget` is not
available in the image.

Co-authored-by: jakubsikorski <jakub.sikorski@intel.com>