ITEP-70408: Region number on scene is misleading (#385)
This pull request updates the labeling logic for ROIs and tripwires in the sscape.js file to use user-provided titles instead of sequential numbers. Now, if a title is set in the corresponding form input, it will be shown as the label; otherwise, the label is removed.

Labeling logic improvements:

Updated numberRois() to display the ROI title from the input field as the label, or remove the label if no title is set. The label is no longer a sequential number.
Updated numberTripwires() to display the tripwire title from the input field as the label, or remove the label if no title is set. The label is no longer a sequential number.

Fixes #63.