ITEP-74968: Region of View is properly computed (#400)
Fixes #314 by leveraging horizon culling when the camera region of view has points that do not intersect with the XY plane.

Why horizon culling:
Rays on earth that extend from camera will not be able to see beyond the earth's curvature. So we clamp at that distance. When we start using SceneScape on other planets, we will make the curvature value configurable.