ITEP-80706: Align the generated mesh in the first quadrant (#575)
This PR aligns the generated mesh from mapping service to be in the first quadrant (+x and +y) and the rotation of the mesh is adjusted to have the floor plane flatter on the xy plane. The ortho/top-down view of the generated mesh i.e the thumbnail is properly rendered. The pose of all cameras is accurately reflected in scenescape.